+++ common/src/main/java/com/gitlab/srcmc/rctapi/api/ai/utils/MoveType.java
@@ -19,7 +19,6 @@ package com.gitlab.srcmc.rctapi.api.ai.utils;
 
 import java.util.Map;
 import com.cobblemon.mod.common.api.moves.categories.DamageCategories;
-import com.cobblemon.mod.common.api.types.ElementalTypes;
 import com.cobblemon.mod.common.battles.InBattleMove;
 import com.cobblemon.mod.common.battles.pokemon.BattlePokemon;
 
@@ -51,6 +50,18 @@ public enum MoveType {
         MOVE_HANDLERS.getOrDefault(moveId, (f, t) -> {}).handle(from, to);
     }
 
+    private static boolean ability(BattlePokemon p, String... names) {
+        var a = p.getEffectedPokemon().getAbility().getName();
+
+        for(var n : names) {
+            if(a.equals(n)) {
+                return true;
+            }
+        }
+
+        return false;
+    }
+
     // custom move evaluations
     private static final Map<String, Evaluator> MOVE_EVALUATORS = Map.<String, Evaluator>ofEntries(
         Map.<String, Evaluator>entry("attract", (from, to, move) -> BattleEffects.Pokemon.Volatile.attract(to) ? 0.0 : 1.0),
@@ -60,14 +71,14 @@ public enum MoveType {
         Map.<String, Evaluator>entry("swagger", (from, to, move) -> (BattleEffects.Pokemon.Volatile.confusion(to) || (!BattleEffects.Pokemon.State.raised(to) && BattleEffects.Field.Terrain.mistyterrain(to))) ? 0.0 : 1.0),
         Map.<String, Evaluator>entry("sweetkiss", (from, to, move) -> (BattleEffects.Pokemon.Volatile.confusion(to) || (!BattleEffects.Pokemon.State.raised(to) && BattleEffects.Field.Terrain.mistyterrain(to))) ? 0.0 : 1.0),
         Map.<String, Evaluator>entry("teeterdance", (from, to, move) -> (BattleEffects.Pokemon.Volatile.confusion(to) || (!BattleEffects.Pokemon.State.raised(to) && BattleEffects.Field.Terrain.mistyterrain(to))) ? 0.0 : 1.0),
-        Map.<String, Evaluator>entry("curse", (from, to, move) -> ((ElementalTypes.INSTANCE.getGHOST().equals(from.getEffectedPokemon().getPrimaryType()) || ElementalTypes.INSTANCE.getGHOST().equals(from.getEffectedPokemon().getSecondaryType())) && BattleEffects.Pokemon.Volatile.cursed(to)) ? 0.0 : 1.0),
+        Map.<String, Evaluator>entry("curse", (from, to, move) -> (TypeChart.is(from, TypeChart.GHOST) && BattleEffects.Pokemon.Volatile.cursed(to)) ? 0.0 : 1.0),
         Map.<String, Evaluator>entry("leechseed", (from, to, move) -> BattleEffects.Pokemon.Volatile.leech(to) ? 0.0 : 1.0),
         Map.<String, Evaluator>entry("ingrain", (from, to, move) -> BattleEffects.Pokemon.Volatile.ingrain(from) ? 0.0 : 1.0),
         Map.<String, Evaluator>entry("aquaring", (from, to, move) -> BattleEffects.Pokemon.Volatile.aquaring(from) ? 0.0 : 1.0),
         // Map.<String, Evaluator>entry("smackdown", (from, to, move) -> BattleEffects.Pokemon.Volatile.smackdown(to) ? 0.0 : 1.0),
         Map.<String, Evaluator>entry("telekinesis", (from, to, move) -> BattleEffects.Pokemon.State.raised(to) ? 0.0 : 1.0),
         Map.<String, Evaluator>entry("magnetrise", (from, to, move) -> BattleEffects.Pokemon.State.raised(from) ? 0.0 : 1.0),
-        Map.<String, Evaluator>entry("flowershield", (from, to, move) -> from.getActor().getSide().getActivePokemon().stream().anyMatch(p -> ElementalTypes.INSTANCE.getGRASS().equals(p.getBattlePokemon().getEffectedPokemon().getPrimaryType()) || ElementalTypes.INSTANCE.getGRASS().equals(p.getBattlePokemon().getEffectedPokemon().getSecondaryType())) ? 1.0 : 0.0),
+        Map.<String, Evaluator>entry("flowershield", (from, to, move) -> from.getActor().getSide().getActivePokemon().stream().filter(p -> p.hasPokemon()).map(p -> p.getBattlePokemon()).anyMatch(p -> TypeChart.is(p, TypeChart.GRASS)) ? 1.0 : 0.0),
         Map.<String, Evaluator>entry("auroraveil", (from, to, move) -> (!BattleEffects.Side.Screen.auroraveil(from) && (BattleEffects.Field.Weather.snow(from) || BattleEffects.Field.Weather.hail(from))) ? 1.0 : 0.0),
         Map.<String, Evaluator>entry("reflect", (from, to, move) -> (BattleEffects.Side.Screen.reflect(from) || BattleEffects.Side.Screen.auroraveil(from)) ? 0.0 : 1.0),
         Map.<String, Evaluator>entry("lightscreen", (from, to, move) -> (BattleEffects.Side.Screen.lightscreen(from) || BattleEffects.Side.Screen.auroraveil(from)) ? 0.0 : 1.0),
@@ -79,21 +90,42 @@ public enum MoveType {
         Map.<String, Evaluator>entry("grassyterrain", (from, to, move) -> BattleEffects.Field.Terrain.grassyterrain(from) ? 0.0 : 1.0),
         Map.<String, Evaluator>entry("mistyterrain", (from, to, move) -> BattleEffects.Field.Terrain.mistyterrain(from) ? 0.0 : 1.0),
         Map.<String, Evaluator>entry("psychicterrain", (from, to, move) -> BattleEffects.Field.Terrain.psychicterrain(from) ? 0.0 : 1.0),
-        Map.<String, Evaluator>entry("sleeptalk", (from, to, move) -> BattleEffects.Pokemon.Status.slp(to) ? 1.0 : 0.0),
-        Map.<String, Evaluator>entry("yawn", (from, to, move) -> (BattleEffects.Pokemon.Volatile.yawn(to) || BattleEffects.Pokemon.Status.slp(to)) ? 0.0 : 1.0),
-        Map.<String, Evaluator>entry("wish", (from, to, move) -> BattleStates.get(from.getActor().battle).getPokemonState(from).has(BattleEffects.Custom.WISH) ? 0.0 : 1.0),
+        Map.<String, Evaluator>entry("sleeptalk", (from, to, move) -> (BattleEffects.Pokemon.Status.slp(from) || ability(from, "comatose")) ? 1.0 : 0.0),
+        Map.<String, Evaluator>entry("wish", (from, to, move) -> BattleStates.get(from.actor.battle).getPokemonState(from).has(BattleEffects.Custom.WISH) ? 0.0 : 1.0),
         Map.<String, Evaluator>entry("meanlook", (from, to, move) -> BattleEffects.Pokemon.State.trapped(to) ? 0.0 : 1.0),
         Map.<String, Evaluator>entry("spiderweb", (from, to, move) -> BattleEffects.Pokemon.State.trapped(to) ? 0.0 : 1.0),
         Map.<String, Evaluator>entry("block", (from, to, move) -> BattleEffects.Pokemon.State.trapped(to) ? 0.0 : 1.0),
-        Map.<String, Evaluator>entry("fakeout", (from, to, move) -> BattleStates.get(from.actor.battle).getPokemonState(from).age(BattleEffects.Custom.TURN) > 1 ? 0.0 : 1.0),
+        Map.<String, Evaluator>entry("fakeout", (from, to, move) -> BattleStates.get(from.actor.battle).getPokemonState(from).age(BattleEffects.Custom.TURN) > 1 ? 0.0 : 1.25),
         Map.<String, Evaluator>entry("tailwind", (from, to, move) -> BattleEffects.Side.Tailwind.tailwind(from) ? 0.0 : 1.0),
         Map.<String, Evaluator>entry("gravity", (from, to, move) -> BattleEffects.Field.Gravity.gravity(from) ? 0.0 : 1.0),
         Map.<String, Evaluator>entry("trickroom", (from, to, move) -> BattleEffects.Field.Room.trickroom(from) ? 0.0 : 1.0),
         Map.<String, Evaluator>entry("spikes", (from, to, move) -> (3 - BattleEffects.Side.Hazard.spikes(to))/3.0),
         Map.<String, Evaluator>entry("stealthrock", (from, to, move) -> BattleEffects.Side.Hazard.stealthrock(to) > 0 ? 0.0 : 1.0),
         Map.<String, Evaluator>entry("toxicspikes", (from, to, move) -> (2 - BattleEffects.Side.Hazard.toxicspikes(to))/2.0),
-        Map.<String, Evaluator>entry("stickyweb", (from, to, move) -> BattleEffects.Side.Hazard.stickyweb(to) > 0 ? 0.0 : 1.0)
-        // Map.<String, Evaluator>entry("spikes", (from, to, move) -> BattleEffects.Side.Hazard.sharpsteel(to) > 0 ? 0.0 : 1.0),
+        Map.<String, Evaluator>entry("stickyweb", (from, to, move) -> BattleEffects.Side.Hazard.stickyweb(to) > 0 ? 0.0 : 1.0),
+        // Map.<String, Evaluator>entry("steelsurge", (from, to, move) -> BattleEffects.Side.Hazard.sharpsteel(to) > 0 ? 0.0 : 1.0),
+        Map.<String, Evaluator>entry("protect", (from, to, move) -> BattleStates.get(from.actor.battle).getPokemonState(from).has(BattleEffects.Custom.PROTECT) ? 0.15 : 1.0),
+        Map.<String, Evaluator>entry("taunt", (from, to, move) -> BattleEffects.Pokemon.Volatile.taunt(to) ? 0 : 1.0),
+        //PARALYSIS https://bulbapedia.bulbagarden.net/wiki/Paralysis_(status_condition)#Prevention TODO: leafguard + harshsunlight, etc.
+        Map.<String, Evaluator>entry("glare", (from, to, move) -> (BattleEffects.Pokemon.Status.any(to) || TypeChart.is(to, TypeChart.ELECTRIC) || ability(to, "limber", "comatose", "purifyingsalt")) ? 0 : 1.0),
+        Map.<String, Evaluator>entry("stunspore", (from, to, move) -> (BattleEffects.Pokemon.Status.any(to) || TypeChart.is(to, TypeChart.ELECTRIC | TypeChart.GRASS) || ability(to, "limber", "comatose", "purifyingsalt")) ? 0 : 1.0),
+        Map.<String, Evaluator>entry("thunderwave", (from, to, move) -> (BattleEffects.Pokemon.Status.any(to) || TypeChart.is(to, TypeChart.ELECTRIC | TypeChart.GROUND) || ability(to, "limber", "comatose", "purifyingsalt")) ? 0 : 1.0),
+        // POISON https://bulbapedia.bulbagarden.net/wiki/Poison_(status_condition)#Prevention TODO: mistyterrain + grounded, etc.
+        Map.<String, Evaluator>entry("poisongas", (from, to, move) -> (BattleEffects.Pokemon.Status.any(to) || (TypeChart.is(to, TypeChart.POISON | TypeChart.STEEL) && !ability(from, "corrosion")) || ability(to, "immunity", "comatose", "purifyingsalt", "pastelveil")) ? 0 : 1.0),
+        Map.<String, Evaluator>entry("poisonpowder", (from, to, move) -> (BattleEffects.Pokemon.Status.any(to) || (TypeChart.is(to, TypeChart.POISON | TypeChart.STEEL | TypeChart.GRASS) && !ability(from, "corrosion")) || ability(to, "immunity", "comatose", "purifyingsalt", "pastelveil")) ? 0 : 1.0),
+        Map.<String, Evaluator>entry("toxicthread", (from, to, move) -> (BattleEffects.Pokemon.Status.any(to) || (TypeChart.is(to, TypeChart.POISON | TypeChart.STEEL) && !ability(from, "corrosion")) || ability(to, "immunity", "comatose", "purifyingsalt", "pastelveil")) ? 0.25 : 1.0),
+        Map.<String, Evaluator>entry("toxic", (from, to, move) -> (BattleEffects.Pokemon.Status.any(to) || (TypeChart.is(to, TypeChart.POISON | TypeChart.STEEL) && !ability(from, "corrosion")) || ability(to, "immunity", "comatose", "purifyingsalt", "pastelveil")) ? 0 : 1.0),
+        // BURN https://bulbapedia.bulbagarden.net/wiki/Burn_(status_condition)#Prevention TODO: leafguard + harshsunlight, etc.
+        Map.<String, Evaluator>entry("willowisp", (from, to, move) -> (BattleEffects.Pokemon.Status.any(to) || TypeChart.is(to, TypeChart.FIRE) || ability(to, "waterveil", "waterbubble", "comatose", "thermalexchange", "purifyingsalt")) ? 0 : 1.0),
+        // SLEEP https://bulbapedia.bulbagarden.net/wiki/Sleep_(status_condition)#Prevention TODO: electricterrain + grounded, etc.
+        Map.<String, Evaluator>entry("darkvoid", (from, to, move) -> (BattleEffects.Pokemon.Status.any(to) || ability(to, "insomnia", "vitalspirit", "purifyingsalt", "comatose")) ? 0.5 : 1.0),
+        Map.<String, Evaluator>entry("grasswhistle", (from, to, move) -> (BattleEffects.Pokemon.Status.any(to) || ability(to, "insomnia", "vitalspirit", "purifyingsalt", "comatose", "soundproof")) ? 0 : 1.0),
+        Map.<String, Evaluator>entry("hypnosis", (from, to, move) -> (BattleEffects.Pokemon.Status.any(to) || ability(to, "insomnia", "vitalspirit", "purifyingsalt", "comatose")) ? 0 : 1.0),
+        Map.<String, Evaluator>entry("lovelykiss", (from, to, move) -> (BattleEffects.Pokemon.Status.any(to) || ability(to, "insomnia", "vitalspirit", "purifyingsalt", "comatose")) ? 0 : 1.0),
+        Map.<String, Evaluator>entry("sing", (from, to, move) -> (BattleEffects.Pokemon.Status.any(to) || ability(to, "insomnia", "vitalspirit", "purifyingsalt", "comatose", "soundproof")) ? 0 : 1.0),
+        Map.<String, Evaluator>entry("sleeppowder", (from, to, move) -> (BattleEffects.Pokemon.Status.any(to) || TypeChart.is(to, TypeChart.GRASS) || ability(to, "insomnia", "vitalspirit", "purifyingsalt", "comatose", "overcoat")) ? 0 : 1.0),
+        Map.<String, Evaluator>entry("spore", (from, to, move) -> (BattleEffects.Pokemon.Status.any(to) || TypeChart.is(to, TypeChart.GRASS) || ability(to, "insomnia", "vitalspirit", "purifyingsalt", "comatose", "overcoat")) ? 0 : 1.0),
+        Map.<String, Evaluator>entry("yawn", (from, to, move) -> (BattleEffects.Pokemon.Volatile.yawn(to) || BattleEffects.Pokemon.Status.any(to) || ability(to, "insomnia", "vitalspirit", "purifyingsalt", "comatose")) ? 0.0 : 1.0)
     );
 
     // custom move event handlers
@@ -101,7 +133,8 @@ public enum MoveType {
         Map.<String, Handler>entry("wish", (f, t) -> BattleStates.get(f.actor.battle).getPokemonState(f).add(BattleEffects.Custom.WISH)),
         Map.<String, Handler>entry("meanlook", (f, t) -> BattleStates.get(f.actor.battle).getPokemonState(f).add(BattleEffects.Custom.MEANLOOK)),
         Map.<String, Handler>entry("spiderweb", (f, t) -> BattleStates.get(f.actor.battle).getPokemonState(f).add(BattleEffects.Custom.SPIDERWEB)),
-        Map.<String, Handler>entry("block", (f, t) -> BattleStates.get(f.actor.battle).getPokemonState(f).add(BattleEffects.Custom.BLOCK))
+        Map.<String, Handler>entry("block", (f, t) -> BattleStates.get(f.actor.battle).getPokemonState(f).add(BattleEffects.Custom.BLOCK)),
+        Map.<String, Handler>entry("protect", (f, t) -> BattleStates.get(f.actor.battle).getPokemonState(f).add(BattleEffects.Custom.PROTECT))
     );
 
     // all moves according to https://pokemondb.net/move/all (05/2025)