+++ CHANGELOG.md
@@ -4,11 +4,16 @@
 
 To update be sure to configure your `GsonBuilder` instances with `RCTApi#configureGsonBuilder(GsonBuilder)` or simply use a builder provided by `RCTApi#gsonBuilder()`.
 
-## [0.12.1-beta] - 2025-05-20
+## [0.12.1-beta] - 2025-05-21
 
 **Changed:**
 
-- *#66* Improved awareness of RCTBattleAI for certain moves (tailwind, gravity, trickroom, spikes, stealthrock, toxicspikes and stickyweb)
+- *#66* General improvements for `RCTBattleAI`
+  - Custom evaluations for hazard and field manipulating moves (tailwind, gravity, trickroom, spikes, stealthrock, toxicspikes and stickyweb)
+  - Custom evaluations for moves that inflict major status conditions (like spore or glare)
+  - Custom evaluations for some common moves with very specific effects (like taunt or protect)
+  - Custom evaluations for weather and terrain moves (like raindance or mistyterrain)
+  - Some minor fixes and adjustments here and there (e.g. sleeptalk checking the opponents sleep status instead of the users)
 - *#65* Refactored `BattleEffects` (former `PokeContext`) utility class for the battle ai
 
 **Fixed:**